wk9-resnetaws/requirements.txt
# Core PyTorch and Vision
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data and ML
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.3.0
Pillow>=9.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Progress bars and utilities
tqdm>=4.64.0

# AWS SDK
boto3>=1.26.0
botocore>=1.29.0

# System utilities
psutil>=5.9.0

# Optional: For ONNX export
# onnx>=1.14.0
# onnxruntime>=1.15.0

# Optional: For model profiling
# thop>=0.1.1

# Optional: For TensorBoard logging
# tensorboard>=2.13.0

# Development and testing (optional)
# pytest>=7.0.0
# black>=23.0.0
# flake8>=6.0.0

# Jupyter notebook support
jupyter>=1.0.0
ipywidgets>=8.0.0